+++ code/graph_algorithms/src/bipartite_checking/bipartite_checking_adjacency_list.java
